+++ app/src/main/java/com/example/taskyapplication/agenda/di/DatabaseModule.kt
@@ -1,0 +1,35 @@
+package com.example.taskyapplication.agenda.di
+
+import android.content.Context
+import androidx.room.Room
+import com.example.taskyapplication.agenda.task.data.local.dao.TaskDao
+import com.example.taskyapplication.agenda.task.data.local.dao.TaskDatabase
+import dagger.Module
+import dagger.Provides
+import dagger.hilt.InstallIn
+import dagger.hilt.android.qualifiers.ApplicationContext
+import dagger.hilt.components.SingletonComponent
+import javax.inject.Singleton
+
+@InstallIn(SingletonComponent::class)
+@Module
+object DatabaseModule {
+
+    @Provides
+    fun provideTaskDao(database: TaskDatabase): TaskDao {
+        return database.taskDao()
+    }
+
+    @Provides
+    @Singleton
+    fun provideTaskDatabase(@ApplicationContext context: Context): TaskDatabase {
+        return Room.databaseBuilder(
+                context = context,
+                klass = TaskDatabase::class.java,
+                name = "task_database"
+            ).fallbackToDestructiveMigration(false)
+            .build()
+    }
+
+
+}
