+++ app/src/main/java/com/example/taskyapplication/agenda/task/data/local/dao/TaskDatabase.kt
@@ -1,9 +1,10 @@
 package com.example.taskyapplication.agenda.task.data.local.dao
 
 import androidx.room.Database
-import com.example.taskyapplication.agenda.task.data.local.entities.TaskEntity
+import androidx.room.RoomDatabase
+import com.example.taskyapplication.agenda.task.data.local.entity.TaskEntity
 
 @Database(entities = [TaskEntity::class], version = 1)
-abstract class TaskDatabase {
+abstract class TaskDatabase : RoomDatabase() {
     abstract fun taskDao(): TaskDao
 }
