+++ app/src/main/java/com/example/taskyapplication/agenda/task/data/local/entity/TaskEntity.kt
@@ -1,9 +1,8 @@
-package com.example.taskyapplication.agenda.task.data.local.entities
+package com.example.taskyapplication.agenda.task.data.local.entity
 
 import androidx.room.ColumnInfo
 import androidx.room.Entity
 import androidx.room.PrimaryKey
-import com.example.taskyapplication.agenda.task.domain.model.TaskExternalModel
 
 @Entity(tableName = "tasks")
 data class TaskEntity(
@@ -14,13 +13,4 @@ data class TaskEntity(
     val time: Long,
     @ColumnInfo(name = "remind_at") val remindAt: Long,
     val isDone: Boolean
-)
-
-fun TaskEntity.asTask() = TaskExternalModel(
-    id = id,
-    title = title,
-    description = description,
-    time = time,
-    remindAt = remindAt,
-    isDone = isDone
 )
