+++ app/src/main/java/com/example/taskyapplication/agenda/task/data/mappers/ModelMappers.kt
@@ -1,0 +1,24 @@
+package com.example.taskyapplication.agenda.task.data.mappers
+
+import com.example.taskyapplication.agenda.task.data.local.entity.TaskEntity
+import com.example.taskyapplication.agenda.task.data.network.models.TaskNetworkModel
+import com.example.taskyapplication.agenda.task.domain.model.TaskUiState
+
+
+fun TaskNetworkModel.asEntity() = TaskEntity(
+    id = eventId,
+    title = title,
+    description = description,
+    time = startTime,
+    remindAt = reminderTime,
+    isDone = isDone
+)
+
+fun TaskEntity.asTask() = TaskUiState(
+    id = id,
+    title = title,
+    description = description,
+    time = time,
+    remindAt = remindAt,
+    isDone = isDone
+)
