+++ app/src/main/java/com/example/taskyapplication/agenda/task/domain/model/TaskUiState.kt
@@ -1,6 +1,6 @@
 package com.example.taskyapplication.agenda.task.domain.model
 
-data class TaskExternalModel(
+data class TaskUiState(
     val id: String,
     val title: String,
     val description: String,
